+++ src/filetype.h
@@ -16,7 +16,7 @@ struct file_type {
   const char* (*name)(void);
   const char* (*type)(struct file_type* base);
   int (*mark)(struct document_text_render_info* render_info);
-  int (*bracket_match)(struct document_text_render_info* render_info);
+  size_t (*bracket_match)(struct document_text_render_info* render_info);
 
   struct config* config;
   char* file_type;
@@ -31,7 +31,7 @@ struct file_type {
 
 struct trie_node* file_type_config_base(struct file_type* base, const char* suffix);
 int file_type_keyword_config(struct file_type* base, struct encoding_cache* cache, struct trie_node* parent, int* keyword_length, int nocase);
-int file_type_bracket_match(struct document_text_render_info* render_info);
+size_t file_type_bracket_match(struct document_text_render_info* render_info);
 const char* file_type_file_type(struct file_type* base);
 
 #endif  /* #ifndef TIPPSE_FILETYPE_H */
